fix(security): tighten input parsing in audit queries, cursors and the seal threshold
Audit log queries check sort_by and sort_order against a fixed column list
before building the ORDER BY clause. Integer parsing in pagination cursors,
query limits and local user lookup uses explicit bit sizes so out-of-range
input is rejected at the parse step. The unseal cooldown threshold is typed
as uint32; a malformed value refuses startup rather than falling back to the
default.

AI-assisted: Claude (Anthropic); Codex GPT-5.6 (OpenAI), independent review

+++ backend/cmd/server/sealwire.go
@@ -111,7 +111,7 @@ func newSealMachine(s1 *stage1, swap *swappableHandler) (*sealWiring, error) {
 	limiter := seal.NewLimiter(seal.LimiterConfig{
 		BaseBackoff:       sealCfg.BackoffBase,
 		MaxBackoff:        sealCfg.BackoffMax,
-		GlobalThreshold:   uint32(sealCfg.CooldownThreshold),
+		GlobalThreshold:   sealCfg.CooldownThreshold,
 		GlobalCooldown:    sealCfg.Cooldown,
 		MaxGlobalCooldown: sealCfg.CooldownMax,
 	})
